requirements.txt
pycups
loguru
argparse
systemd-python
